refactor(batch-b): split actions/models/skills hotspots into focused submodules
Apply the thin-facade pattern across the next wave of Batch B modules,
keeping each file under ~500 lines and carving stable seams by
responsibility. Public API surface and behaviour are unchanged.

dcc-mcp-actions:
  - registry/mod.rs now re-exports from
    registry/meta.rs (ActionMeta struct and defaults) and
    registry/python.rs (PyO3 #[pymethods] impl).

dcc-mcp-models:
  - skill_metadata/mod.rs re-exports from
    skill_metadata/methods.rs (inherent impls),
    skill_metadata/serde_impl.rs (deserialize helpers), and
    skill_metadata/python.rs (PyO3 bindings).
  - skill_metadata/tool_declaration.rs extracts its PyO3 surface for
    SkillGroup and ToolDeclaration into tool_declaration_python.rs.

dcc-mcp-skills:
  - catalog/catalog.rs delegates discovery/construction to
    catalog/catalog_discovery.rs and load/unload lifecycle to
    catalog/catalog_loading.rs.
  - loader/mod.rs becomes a thin facade re-exporting from
    loader/files.rs (fs enumeration helpers) and
    loader/scan.rs (scan_and_load, scan_and_load_lenient, and the
    crate-internal load_all_skills helper gated to #[cfg(test)]).
  - validator.rs now wires in
    validator_types.rs (IssueSeverity/IssueCategory/SkillValidationIssue/
    SkillValidationReport), validator_rules.rs (all validate_* rules),
    validator_python.rs (PyO3 bindings), and
    validator_tests.rs (unit tests), via #[path] module inclusion.

docs:
  - Architecture guide gains "Maintainer layout" notes for the new
    skill_metadata/tool_declaration, loader, and validator splits so
    agents can locate the right file before editing.
