+++ crates/dcc-mcp-actions/src/registry/mod.rs
@@ -7,8 +7,6 @@ use pyo3::prelude::*;
 use dcc_mcp_utils::py_json::json_value_to_pyobject;
 
 use dashmap::DashMap;
-use dcc_mcp_models::{ExecutionMode, NextTools, ThreadAffinity, ToolAnnotations};
-use serde::{Deserialize, Serialize};
 use std::sync::Arc;
 
 #[cfg(feature = "python-bindings")]
@@ -17,134 +15,11 @@ use dcc_mcp_utils::constants::{DEFAULT_DCC, DEFAULT_VERSION};
 #[cfg(feature = "python-bindings")]
 use dcc_mcp_utils::constants::default_schema;
 
-/// Metadata about a registered Action (stored in Rust).
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
-#[serde(default)]
-pub struct ActionMeta {
-    /// Unique action identifier.
-    pub name: String,
-    /// Human-readable action description.
-    pub description: String,
-    /// Action category for grouping (e.g. "geometry", "pipeline").
-    pub category: String,
-    /// Searchable tags for discovery.
-    pub tags: Vec<String>,
-    /// Target DCC application (e.g. "maya", "blender").
-    pub dcc: String,
-    /// Semantic version string.
-    pub version: String,
-    /// JSON Schema for action input parameters.
-    pub input_schema: serde_json::Value,
-    /// JSON Schema for action output.
-    pub output_schema: serde_json::Value,
-    /// Optional path to the Python source file defining this action.
-    pub source_file: Option<String>,
-    /// Name of the skill this action belongs to (if registered from a skill).
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub skill_name: Option<String>,
-    /// Tool group this action belongs to (``""`` = always active).
-    ///
-    /// See [`dcc_mcp_models::SkillGroup`]; used together with `enabled` to
-    /// implement progressive tool exposure via ``activate_tool_group``.
-    #[serde(default)]
-    pub group: String,
-    /// Whether this action is currently active / callable.
-    ///
-    /// Tools in an inactive group are collapsed into a ``__group__<name>``
-    /// stub in ``tools/list``. The dispatcher refuses to invoke disabled
-    /// actions.
-    #[serde(default = "default_enabled")]
-    pub enabled: bool,
-    /// Host-DCC capabilities required for this action to be surfaced.
-    ///
-    /// When non-empty, Gateway / adapter implementations **should** hide
-    /// this action from ``tools/list`` on sessions whose host DCC does not
-    /// advertise every listed capability (see
-    /// [``WebViewAdapter.capabilities``](crate::adapters::webview) for the
-    /// pre-defined key set: ``"scene"``, ``"timeline"``, ``"selection"``,
-    /// ``"undo"``, ``"render"``).
-    ///
-    /// The registry itself does **not** perform filtering — filtering is
-    /// the responsibility of the consumer (Gateway, HTTP server, adapter).
-    /// Storing the declaration here avoids a separate side-table lookup.
-    #[serde(default, skip_serializing_if = "Vec::is_empty")]
-    pub required_capabilities: Vec<String>,
-    /// Execution mode declared by the skill author (issue #317).
-    ///
-    /// `Sync` (default) or `Async`. Drives the server-derived MCP
-    /// `deferredHint` annotation emitted by `tools/list`.
-    #[serde(default)]
-    pub execution: ExecutionMode,
-    /// Optional hint about typical execution time in seconds (issue #317).
-    ///
-    /// Surfaces under `_meta.dcc.timeoutHintSecs` on the tool definition —
-    /// never inside `annotations`.
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub timeout_hint_secs: Option<u32>,
-    /// Thread-affinity hint surfaced by the skill author (issue #332).
-    ///
-    /// Drives async-dispatch routing in the HTTP server:
-    /// `Main` forces the tool through [`crate::DeferredExecutor`] even along
-    /// the async `tools/call` path, guaranteeing the handler runs on the DCC's
-    /// main thread. `Any` (default) allows execution on a Tokio worker.
-    #[serde(default, skip_serializing_if = "is_default_thread_affinity")]
-    pub thread_affinity: ThreadAffinity,
-    /// MCP tool annotations declared by the skill author (issue #344).
-    ///
-    /// When present, each non-`None` hint is surfaced on the MCP
-    /// `tools/list` tool definition as a spec-compliant camelCase field
-    /// (`readOnlyHint`, `destructiveHint`, …). The dcc-mcp-core-specific
-    /// `deferred_hint` lands in `_meta["dcc.deferred_hint"]` rather than
-    /// inside the spec `annotations` map.
-    #[serde(default, skip_serializing_if = "ToolAnnotations::is_empty")]
-    pub annotations: ToolAnnotations,
-    /// Suggested follow-up tools surfaced on `CallToolResult._meta`
-    /// under `dcc.next_tools` (issue #342).
-    ///
-    /// Populated from the per-tool `next-tools` entry in the sibling
-    /// `tools.yaml` file. Tool names must pass
-    /// [`dcc_mcp_naming::validate_tool_name`]; invalid entries are
-    /// dropped at skill-load time with a warning.
-    #[serde(default, skip_serializing_if = "next_tools_is_empty")]
-    pub next_tools: NextTools,
-}
+mod meta;
+#[cfg(feature = "python-bindings")]
+mod python;
 
-fn next_tools_is_empty(nt: &NextTools) -> bool {
-    nt.on_success.is_empty() && nt.on_failure.is_empty()
-}
-
-fn is_default_thread_affinity(affinity: &ThreadAffinity) -> bool {
-    matches!(affinity, ThreadAffinity::Any)
-}
-
-fn default_enabled() -> bool {
-    true
-}
-
-impl Default for ActionMeta {
-    fn default() -> Self {
-        Self {
-            name: String::new(),
-            description: String::new(),
-            category: String::new(),
-            tags: Vec::new(),
-            dcc: String::new(),
-            version: String::new(),
-            input_schema: serde_json::Value::Null,
-            output_schema: serde_json::Value::Null,
-            source_file: None,
-            skill_name: None,
-            group: String::new(),
-            enabled: true,
-            required_capabilities: Vec::new(),
-            execution: ExecutionMode::Sync,
-            timeout_hint_secs: None,
-            thread_affinity: ThreadAffinity::Any,
-            annotations: ToolAnnotations::default(),
-            next_tools: NextTools::default(),
-        }
-    }
-}
+pub use meta::ActionMeta;
 
 /// Thread-safe Action registry.
 ///
@@ -501,458 +376,6 @@ impl ActionRegistry {
         }
         seen
     }
-}
-
-// ── Python bindings ──
-
-#[cfg(feature = "python-bindings")]
-#[pymethods]
-impl ActionRegistry {
-    #[new]
-    fn py_new() -> Self {
-        Self::new()
-    }
-
-    /// Register multiple actions at once from a list of dicts.
-    ///
-    /// Each dict may contain the same keyword arguments as :meth:`register`.
-    /// Fields not present in a dict fall back to their defaults.
-    /// Entries without a ``"name"`` key (or with an empty name) are silently skipped.
-    ///
-    /// Example::
-    ///
-    ///   reg.register_batch([
-    ///       {"name": "create_sphere", "category": "geometry", "dcc": "maya"},
-    ///       {"name": "delete_object", "category": "edit",     "dcc": "maya"},
-    ///   ])
-    #[pyo3(name = "register_batch")]
-    fn py_register_batch(&self, actions: Vec<pyo3::Bound<'_, pyo3::types::PyAny>>) {
-        for item in &actions {
-            let Ok(dict) = item.cast::<pyo3::types::PyDict>() else {
-                continue;
-            };
-            let name: String = dict
-                .get_item("name")
-                .ok()
-                .flatten()
-                .and_then(|v| v.extract().ok())
-                .unwrap_or_default();
-            if name.is_empty() {
-                continue;
-            }
-            let description: String = dict
-                .get_item("description")
-                .ok()
-                .flatten()
-                .and_then(|v| v.extract().ok())
-                .unwrap_or_default();
-            let category: String = dict
-                .get_item("category")
-                .ok()
-                .flatten()
-                .and_then(|v| v.extract().ok())
-                .unwrap_or_default();
-            let tags: Vec<String> = dict
-                .get_item("tags")
-                .ok()
-                .flatten()
-                .and_then(|v| v.extract().ok())
-                .unwrap_or_default();
-            let dcc: String = dict
-                .get_item("dcc")
-                .ok()
-                .flatten()
-                .and_then(|v| v.extract().ok())
-                .unwrap_or_else(|| DEFAULT_DCC.to_string());
-            let version: String = dict
-                .get_item("version")
-                .ok()
-                .flatten()
-                .and_then(|v| v.extract().ok())
-                .unwrap_or_else(|| DEFAULT_VERSION.to_string());
-            let input_schema_str: Option<String> = dict
-                .get_item("input_schema")
-                .ok()
-                .flatten()
-                .and_then(|v| v.extract().ok());
-            let output_schema_str: Option<String> = dict
-                .get_item("output_schema")
-                .ok()
-                .flatten()
-                .and_then(|v| v.extract().ok());
-            let source_file: Option<String> = dict
-                .get_item("source_file")
-                .ok()
-                .flatten()
-                .and_then(|v| v.extract().ok());
-            let skill_name: Option<String> = dict
-                .get_item("skill_name")
-                .ok()
-                .flatten()
-                .and_then(|v| v.extract().ok());
-            let group: String = dict
-                .get_item("group")
-                .ok()
-                .flatten()
-                .and_then(|v| v.extract().ok())
-                .unwrap_or_default();
-            let enabled: bool = dict
-                .get_item("enabled")
-                .ok()
-                .flatten()
-                .and_then(|v| v.extract().ok())
-                .unwrap_or(true);
-            let required_capabilities: Vec<String> = dict
-                .get_item("required_capabilities")
-                .ok()
-                .flatten()
-                .and_then(|v| v.extract().ok())
-                .unwrap_or_default();
-            let execution_str: Option<String> = dict
-                .get_item("execution")
-                .ok()
-                .flatten()
-                .and_then(|v| v.extract().ok());
-            let execution = match execution_str.as_deref() {
-                None | Some("sync") => ExecutionMode::Sync,
-                Some("async") => ExecutionMode::Async,
-                Some(other) => {
-                    tracing::warn!(
-                        "Invalid execution mode {other:?} for '{name}' — defaulting to sync"
-                    );
-                    ExecutionMode::Sync
-                }
-            };
-            let timeout_hint_secs: Option<u32> = dict
-                .get_item("timeout_hint_secs")
-                .ok()
-                .flatten()
-                .and_then(|v| v.extract().ok());
-            let thread_affinity_str: Option<String> = dict
-                .get_item("thread_affinity")
-                .ok()
-                .flatten()
-                .and_then(|v| v.extract().ok());
-            let thread_affinity = match thread_affinity_str.as_deref() {
-                None => ThreadAffinity::Any,
-                Some(s) => match ThreadAffinity::parse(s) {
-                    Some(a) => a,
-                    None => {
-                        tracing::warn!(
-                            "Invalid thread_affinity {s:?} for '{name}' — defaulting to 'any'"
-                        );
-                        ThreadAffinity::Any
-                    }
-                },
-            };
-
-            let input_schema =
-                parse_schema_or_default(input_schema_str.as_deref(), "input_schema", &name);
-            let output_schema =
-                parse_schema_or_default(output_schema_str.as_deref(), "output_schema", &name);
-
-            self.register_action(ActionMeta {
-                name,
-                description,
-                category,
-                tags,
-                dcc,
-                version,
-                input_schema,
-                output_schema,
-                source_file,
-                skill_name,
-                group,
-                enabled,
-                required_capabilities,
-                execution,
-                timeout_hint_secs,
-                thread_affinity,
-                annotations: ToolAnnotations::default(),
-                next_tools: NextTools::default(),
-            });
-        }
-    }
-
-    /// Unregister an action by name.
-    ///
-    /// If ``dcc_name`` is ``None`` (default), the action is removed from the global
-    /// registry and every per-DCC map that contains it.
-    ///
-    /// If ``dcc_name`` is provided, only the per-DCC entry is removed; the global
-    /// entry is removed only when no other DCC still references the action.
-    ///
-    /// Returns ``True`` if the action was found and removed, ``False`` otherwise.
-    ///
-    /// Example::
-    ///
-    ///   reg.register(name="create_sphere", dcc="maya")
-    ///   assert reg.unregister("create_sphere") is True
-    ///   assert reg.unregister("create_sphere") is False  # already gone
-    #[pyo3(name = "unregister")]
-    #[pyo3(signature = (name, dcc_name=None))]
-    fn py_unregister(&self, name: &str, dcc_name: Option<&str>) -> bool {
-        self.unregister(name, dcc_name)
-    }
-
-    /// Register an action.
-    #[allow(clippy::too_many_arguments)]
-    #[pyo3(signature = (name, description="".to_string(), category="".to_string(), tags=vec![], dcc=DEFAULT_DCC.to_string(), version=DEFAULT_VERSION.to_string(), input_schema=None, output_schema=None, source_file=None, skill_name=None, group="".to_string(), enabled=true, required_capabilities=None, execution="sync".to_string(), timeout_hint_secs=None, thread_affinity="any".to_string()))]
-    fn register(
-        &self,
-        name: String,
-        description: String,
-        category: String,
-        tags: Vec<String>,
-        dcc: String,
-        version: String,
-        input_schema: Option<String>,
-        output_schema: Option<String>,
-        source_file: Option<String>,
-        skill_name: Option<String>,
-        group: String,
-        enabled: bool,
-        required_capabilities: Option<Vec<String>>,
-        execution: String,
-        timeout_hint_secs: Option<u32>,
-        thread_affinity: String,
-    ) -> pyo3::PyResult<()> {
-        let input_schema = parse_schema_or_default(input_schema.as_deref(), "input_schema", &name);
-        let output_schema =
-            parse_schema_or_default(output_schema.as_deref(), "output_schema", &name);
-        let execution = match execution.as_str() {
-            "sync" => ExecutionMode::Sync,
-            "async" => ExecutionMode::Async,
-            other => {
-                return Err(pyo3::exceptions::PyValueError::new_err(format!(
-                    "execution must be 'sync' or 'async' (got {other:?})",
-                )));
-            }
-        };
-        let thread_affinity = ThreadAffinity::parse(&thread_affinity).ok_or_else(|| {
-            pyo3::exceptions::PyValueError::new_err(format!(
-                "thread_affinity must be 'any' or 'main' (got {thread_affinity:?})"
-            ))
-        })?;
-
-        self.register_action(ActionMeta {
-            name,
-            description,
-            category,
-            tags,
-            dcc,
-            version,
-            input_schema,
-            output_schema,
-            source_file,
-            skill_name,
-            group,
-            enabled,
-            required_capabilities: required_capabilities.unwrap_or_default(),
-            execution,
-            timeout_hint_secs,
-            thread_affinity,
-            annotations: ToolAnnotations::default(),
-            next_tools: NextTools::default(),
-        });
-        Ok(())
-    }
-
-    /// Enable or disable every action belonging to ``group``.
-    ///
-    /// Returns the number of actions whose ``enabled`` flag changed.
-    #[pyo3(name = "set_group_enabled")]
-    fn py_set_group_enabled(&self, group: &str, enabled: bool) -> usize {
-        self.set_group_enabled(group, enabled)
-    }
-
-    /// Enable or disable a single action.
-    ///
-    /// Returns ``True`` if the action existed.
-    #[pyo3(name = "set_action_enabled")]
-    fn py_set_action_enabled(&self, name: &str, enabled: bool) -> bool {
-        self.set_action_enabled(name, enabled)
-    }
-
-    /// List all currently-enabled actions.
-    #[pyo3(name = "list_actions_enabled")]
-    #[pyo3(signature = (dcc_name=None))]
-    fn py_list_actions_enabled(
-        &self,
-        py: Python,
-        dcc_name: Option<&str>,
-    ) -> PyResult<Vec<Py<PyAny>>> {
-        self.list_actions_enabled(dcc_name)
-            .iter()
-            .map(|meta| action_meta_to_py(py, meta))
-            .collect()
-    }
-
-    /// List actions belonging to ``group`` (across all DCCs).
-    #[pyo3(name = "list_actions_in_group")]
-    fn py_list_actions_in_group(&self, py: Python, group: &str) -> PyResult<Vec<Py<PyAny>>> {
-        self.list_actions_in_group(group)
-            .iter()
-            .map(|meta| action_meta_to_py(py, meta))
-            .collect()
-    }
-
-    /// Return distinct non-empty group names.
-    #[pyo3(name = "list_groups")]
-    fn py_list_groups(&self) -> Vec<String> {
-        self.list_groups()
-    }
-
-    /// Get action metadata as dict.
-    #[pyo3(name = "get_action")]
-    #[pyo3(signature = (name, dcc_name=None))]
-    fn py_get_action(
-        &self,
-        py: Python,
-        name: &str,
-        dcc_name: Option<&str>,
-    ) -> PyResult<Option<Py<PyAny>>> {
-        self.get_action(name, dcc_name)
-            .map(|meta| action_meta_to_py(py, &meta))
-            .transpose()
-    }
-
-    /// List all action names for a DCC.
-    #[pyo3(name = "list_actions_for_dcc")]
-    fn py_list_actions_for_dcc(&self, dcc_name: &str) -> Vec<String> {
-        self.list_actions_for_dcc(dcc_name)
-    }
-
-    /// List all actions with metadata.
-    #[pyo3(name = "list_actions")]
-    #[pyo3(signature = (dcc_name=None))]
-    fn py_list_actions(&self, py: Python, dcc_name: Option<&str>) -> PyResult<Vec<Py<PyAny>>> {
-        self.list_actions(dcc_name)
-            .iter()
-            .map(|meta| action_meta_to_py(py, meta))
-            .collect()
-    }
-
-    /// Get all registered DCC names.
-    #[pyo3(name = "get_all_dccs")]
-    fn py_get_all_dccs(&self) -> Vec<String> {
-        self.get_all_dccs()
-    }
-
-    /// Search actions by category, tags, and/or DCC name.
-    ///
-    /// All provided filters are AND-ed together:
-    ///
-    /// - ``category``: exact category match (``None`` or empty string = no filter)
-    /// - ``tags``: action must contain **all** listed tags (empty list = no filter)
-    /// - ``dcc_name``: scoped to a specific DCC (``None`` = all DCCs)
-    ///
-    /// Returns a list of action metadata dicts.
-    ///
-    /// Example::
-    ///
-    ///   reg.register(name="create_sphere", category="geometry",
-    ///                tags=["create", "mesh"], dcc="maya")
-    ///   results = reg.search_actions(category="geometry", tags=["create"])
-    ///   # [{"name": "create_sphere", ...}]
-    #[pyo3(name = "search_actions")]
-    #[pyo3(signature = (category=None, tags=vec![], dcc_name=None))]
-    fn py_search_actions(
-        &self,
-        py: Python,
-        category: Option<&str>,
-        tags: Vec<String>,
-        dcc_name: Option<&str>,
-    ) -> PyResult<Vec<Py<PyAny>>> {
-        let tag_refs: Vec<&str> = tags.iter().map(String::as_str).collect();
-        self.search_actions(category, &tag_refs, dcc_name)
-            .iter()
-            .map(|meta| action_meta_to_py(py, meta))
-            .collect()
-    }
-
-    /// Get all unique categories in the registry.
-    ///
-    /// Optionally scoped to a specific DCC.
-    #[pyo3(name = "get_categories")]
-    #[pyo3(signature = (dcc_name=None))]
-    fn py_get_categories(&self, dcc_name: Option<&str>) -> Vec<String> {
-        self.get_categories(dcc_name)
-    }
-
-    /// Get all unique tags in the registry.
-    ///
-    /// Optionally scoped to a specific DCC.
-    #[pyo3(name = "get_tags")]
-    #[pyo3(signature = (dcc_name=None))]
-    fn py_get_tags(&self, dcc_name: Option<&str>) -> Vec<String> {
-        self.get_tags(dcc_name)
-    }
-
-    /// Count actions matching the given search criteria.
-    ///
-    /// Convenience wrapper around :meth:`search_actions`.
-    ///
-    /// Example::
-    ///
-    ///   reg.register(name="create_sphere", category="geometry", dcc="maya")
-    ///   assert reg.count_actions(category="geometry") == 1
-    #[pyo3(name = "count_actions")]
-    #[pyo3(signature = (category=None, tags=vec![], dcc_name=None))]
-    fn py_count_actions(
-        &self,
-        category: Option<&str>,
-        tags: Vec<String>,
-        dcc_name: Option<&str>,
-    ) -> usize {
-        let tag_refs: Vec<&str> = tags.iter().map(String::as_str).collect();
-        self.count_actions(category, &tag_refs, dcc_name)
-    }
-
-    /// Reset the registry.
-    #[pyo3(name = "reset")]
-    fn py_reset(&self) {
-        self.reset()
-    }
-
-    fn __len__(&self) -> usize {
-        self.len()
-    }
-
-    fn __contains__(&self, name: &str) -> bool {
-        self.actions.contains_key(name)
-    }
-
-    fn __repr__(&self) -> String {
-        format!("ToolRegistry(actions={})", self.len())
-    }
-}
-
-/// Parse a JSON schema string, falling back to [`default_schema`] on `None` or invalid JSON.
-#[cfg(feature = "python-bindings")]
-fn parse_schema_or_default(
-    json: Option<&str>,
-    field_name: &str,
-    action_name: &str,
-) -> serde_json::Value {
-    match json {
-        Some(s) => serde_json::from_str(s).unwrap_or_else(|e| {
-            tracing::warn!("Invalid {field_name} JSON for '{action_name}': {e} — using default");
-            default_schema().clone()
-        }),
-        None => default_schema().clone(),
-    }
-}
-
-/// Convert [`ActionMeta`] to a Python dict via serde serialization.
-///
-/// This leverages the existing `#[derive(Serialize)]` on `ActionMeta` to avoid
-/// manually enumerating every field — new fields are automatically included.
-#[cfg(feature = "python-bindings")]
-fn action_meta_to_py(py: Python, meta: &ActionMeta) -> PyResult<Py<PyAny>> {
-    let json_val = serde_json::to_value(meta)
-        .map_err(|e| pyo3::exceptions::PyValueError::new_err(e.to_string()))?;
-    json_value_to_pyobject(py, &json_val)
 }
 
 #[cfg(test)]
