+++ crates/dcc-mcp-models/src/skill_metadata/tool_declaration.rs
@@ -10,7 +10,8 @@ use super::{ExecutionMode, ThreadAffinity};
 use pyo3::prelude::*;
 
 #[cfg(feature = "python-bindings")]
-use pyo3::types::{PyAnyMethods, PyDictMethods};
+#[path = "tool_declaration_python.rs"]
+mod python_impl;
 
 // ── ToolAnnotations ───────────────────────────────────────────────────────
 
@@ -485,431 +486,8 @@ pub struct SkillGroup {
     pub default_active: bool,
 }
 
-#[cfg(feature = "python-bindings")]
-#[pymethods]
-impl SkillGroup {
-    #[new]
-    #[pyo3(signature = (name, description="".to_string(), tools=Vec::<String>::new(), default_active=false))]
-    fn new(name: String, description: String, tools: Vec<String>, default_active: bool) -> Self {
-        Self {
-            name,
-            description,
-            tools,
-            default_active,
-        }
-    }
-
-    #[getter]
-    fn name(&self) -> &str {
-        &self.name
-    }
-
-    #[getter]
-    fn description(&self) -> &str {
-        &self.description
-    }
-
-    #[getter]
-    fn tools(&self) -> Vec<String> {
-        self.tools.clone()
-    }
-
-    #[getter]
-    fn default_active(&self) -> bool {
-        self.default_active
-    }
-
-    fn __repr__(&self) -> String {
-        format!(
-            "SkillGroup(name={:?}, tools={}, default_active={})",
-            self.name,
-            self.tools.len(),
-            self.default_active
-        )
-    }
-}
-
 fn is_null_value(v: &serde_json::Value) -> bool {
     v.is_null()
-}
-
-#[cfg(feature = "python-bindings")]
-#[pymethods]
-impl ToolDeclaration {
-    #[new]
-    #[pyo3(signature = (name, description="".to_string(), input_schema=None, output_schema=None, read_only=false, destructive=false, idempotent=false, defer_loading=false, source_file="".to_string(), group="".to_string(), execution="sync".to_string(), timeout_hint_secs=None, thread_affinity="any".to_string()))]
-    #[allow(clippy::too_many_arguments)]
-    fn new(
-        name: String,
-        description: String,
-        input_schema: Option<String>,
-        output_schema: Option<String>,
-        read_only: bool,
-        destructive: bool,
-        idempotent: bool,
-        defer_loading: bool,
-        source_file: String,
-        group: String,
-        execution: String,
-        timeout_hint_secs: Option<u32>,
-        thread_affinity: String,
-    ) -> pyo3::PyResult<Self> {
-        let input_schema = input_schema
-            .and_then(|s| serde_json::from_str(&s).ok())
-            .unwrap_or(serde_json::json!({"type": "object"}));
-        let output_schema = output_schema
-            .and_then(|s| serde_json::from_str(&s).ok())
-            .unwrap_or(serde_json::Value::Null);
-        let execution = match execution.as_str() {
-            "sync" => ExecutionMode::Sync,
-            "async" => ExecutionMode::Async,
-            other => {
-                return Err(pyo3::exceptions::PyValueError::new_err(format!(
-                    "execution must be 'sync' or 'async' (got {other:?})",
-                )));
-            }
-        };
-        let thread_affinity = ThreadAffinity::parse(&thread_affinity).ok_or_else(|| {
-            pyo3::exceptions::PyValueError::new_err(format!(
-                "thread_affinity must be 'any' or 'main' (got {thread_affinity:?})"
-            ))
-        })?;
-        Ok(Self {
-            name,
-            description,
-            input_schema,
-            output_schema,
-            read_only,
-            destructive,
-            idempotent,
-            defer_loading,
-            source_file,
-            next_tools: NextTools::default(),
-            group,
-            execution,
-            timeout_hint_secs,
-            thread_affinity,
-            _deferred_guard: None,
-            annotations: ToolAnnotations::default(),
-            required_capabilities: Vec::new(),
-        })
-    }
-
-    /// Declared DCC capabilities required for this tool (issue #354).
-    ///
-    /// Returns freeform string tags; an empty list means the tool has no
-    /// capability prerequisites beyond what any DCC adapter provides.
-    #[getter]
-    fn required_capabilities(&self) -> Vec<String> {
-        self.required_capabilities.clone()
-    }
-
-    #[setter]
-    fn set_required_capabilities(&mut self, value: Vec<String>) {
-        self.required_capabilities = value;
-    }
-
-    #[getter]
-    fn execution(&self) -> &'static str {
-        match self.execution {
-            ExecutionMode::Sync => "sync",
-            ExecutionMode::Async => "async",
-        }
-    }
-
-    #[setter]
-    fn set_execution(&mut self, value: String) -> pyo3::PyResult<()> {
-        self.execution = match value.as_str() {
-            "sync" => ExecutionMode::Sync,
-            "async" => ExecutionMode::Async,
-            other => {
-                return Err(pyo3::exceptions::PyValueError::new_err(format!(
-                    "execution must be 'sync' or 'async' (got {other:?})",
-                )));
-            }
-        };
-        Ok(())
-    }
-
-    #[getter]
-    fn timeout_hint_secs(&self) -> Option<u32> {
-        self.timeout_hint_secs
-    }
-
-    #[setter]
-    fn set_timeout_hint_secs(&mut self, value: Option<u32>) {
-        self.timeout_hint_secs = value;
-    }
-
-    #[getter]
-    fn group(&self) -> &str {
-        &self.group
-    }
-
-    #[setter]
-    fn set_group(&mut self, value: String) {
-        self.group = value;
-    }
-
-    fn __repr__(&self) -> String {
-        format!("ToolDeclaration(name={:?})", self.name)
-    }
-
-    #[getter]
-    fn name(&self) -> &str {
-        &self.name
-    }
-
-    #[setter]
-    fn set_name(&mut self, value: String) {
-        self.name = value;
-    }
-
-    #[getter]
-    fn description(&self) -> &str {
-        &self.description
-    }
-
-    #[setter]
-    fn set_description(&mut self, value: String) {
-        self.description = value;
-    }
-
-    /// Returns input_schema as a JSON string.
-    #[getter]
-    fn input_schema(&self) -> String {
-        self.input_schema.to_string()
-    }
-
-    /// Set input_schema from a JSON string.
-    #[setter]
-    fn set_input_schema(&mut self, value: String) {
-        self.input_schema =
-            serde_json::from_str(&value).unwrap_or(serde_json::json!({"type": "object"}));
-    }
-
-    /// Returns output_schema as a JSON string (empty string if null).
-    #[getter]
-    fn output_schema(&self) -> String {
-        if self.output_schema.is_null() {
-            String::new()
-        } else {
-            self.output_schema.to_string()
-        }
-    }
-
-    /// Set output_schema from a JSON string.
-    #[setter]
-    fn set_output_schema(&mut self, value: String) {
-        self.output_schema = if value.is_empty() {
-            serde_json::Value::Null
-        } else {
-            serde_json::from_str(&value).unwrap_or(serde_json::Value::Null)
-        };
-    }
-
-    #[getter]
-    fn read_only(&self) -> bool {
-        self.read_only
-    }
-
-    #[setter]
-    fn set_read_only(&mut self, value: bool) {
-        self.read_only = value;
-    }
-
-    #[getter]
-    fn destructive(&self) -> bool {
-        self.destructive
-    }
-
-    #[setter]
-    fn set_destructive(&mut self, value: bool) {
-        self.destructive = value;
-    }
-
-    #[getter]
-    fn idempotent(&self) -> bool {
-        self.idempotent
-    }
-
-    #[setter]
-    fn set_idempotent(&mut self, value: bool) {
-        self.idempotent = value;
-    }
-
-    #[getter]
-    fn defer_loading(&self) -> bool {
-        self.defer_loading
-    }
-
-    #[setter]
-    fn set_defer_loading(&mut self, value: bool) {
-        self.defer_loading = value;
-    }
-
-    #[getter]
-    fn source_file(&self) -> &str {
-        &self.source_file
-    }
-
-    #[setter]
-    fn set_source_file(&mut self, value: String) {
-        self.source_file = value;
-    }
-
-    /// Return the declared MCP tool annotations as a Python dict.
-    ///
-    /// Keys use MCP-spec camelCase (`readOnlyHint`, `destructiveHint`,
-    /// `idempotentHint`, `openWorldHint`). The dcc-mcp-core-specific
-    /// `deferredHint` key is included when set (it lives in `_meta` on
-    /// `tools/list`, but is exposed here for convenience).
-    /// Missing fields are omitted entirely.
-    #[getter]
-    fn annotations(&self, py: pyo3::Python<'_>) -> pyo3::PyResult<Py<PyAny>> {
-        use pyo3::types::PyDict;
-        let d = PyDict::new(py);
-        if let Some(v) = &self.annotations.title {
-            d.set_item("title", v)?;
-        }
-        if let Some(v) = self.annotations.read_only_hint {
-            d.set_item("readOnlyHint", v)?;
-        }
-        if let Some(v) = self.annotations.destructive_hint {
-            d.set_item("destructiveHint", v)?;
-        }
-        if let Some(v) = self.annotations.idempotent_hint {
-            d.set_item("idempotentHint", v)?;
-        }
-        if let Some(v) = self.annotations.open_world_hint {
-            d.set_item("openWorldHint", v)?;
-        }
-        if let Some(v) = self.annotations.deferred_hint {
-            d.set_item("deferredHint", v)?;
-        }
-        Ok(d.into_any().unbind())
-    }
-
-    /// Set the tool annotations from a Python mapping.
-    ///
-    /// Accepts both snake_case (`read_only_hint`) and camelCase
-    /// (`readOnlyHint`) keys.  Unknown keys are ignored.  Pass `None` or
-    /// an empty dict to clear.
-    #[setter]
-    fn set_annotations(
-        &mut self,
-        py: pyo3::Python<'_>,
-        value: Option<Py<PyAny>>,
-    ) -> pyo3::PyResult<()> {
-        use pyo3::types::PyDict;
-        let Some(obj) = value else {
-            self.annotations = ToolAnnotations::default();
-            return Ok(());
-        };
-        let bound = obj.bind(py);
-        if bound.is_none() {
-            self.annotations = ToolAnnotations::default();
-            return Ok(());
-        }
-        let d = bound.cast::<PyDict>().map_err(|_| {
-            pyo3::exceptions::PyTypeError::new_err("annotations must be a dict or None")
-        })?;
-
-        fn get_bool(d: &pyo3::Bound<'_, PyDict>, keys: &[&str]) -> pyo3::PyResult<Option<bool>> {
-            for k in keys {
-                if let Some(v) = d.get_item(k)? {
-                    if v.is_none() {
-                        return Ok(None);
-                    }
-                    return Ok(Some(v.extract::<bool>()?));
-                }
-            }
-            Ok(None)
-        }
-        fn get_str(d: &pyo3::Bound<'_, PyDict>, keys: &[&str]) -> pyo3::PyResult<Option<String>> {
-            for k in keys {
-                if let Some(v) = d.get_item(k)? {
-                    if v.is_none() {
-                        return Ok(None);
-                    }
-                    return Ok(Some(v.extract::<String>()?));
-                }
-            }
-            Ok(None)
-        }
-
-        self.annotations = ToolAnnotations {
-            title: get_str(d, &["title"])?,
-            read_only_hint: get_bool(d, &["read_only_hint", "readOnlyHint"])?,
-            destructive_hint: get_bool(d, &["destructive_hint", "destructiveHint"])?,
-            idempotent_hint: get_bool(d, &["idempotent_hint", "idempotentHint"])?,
-            open_world_hint: get_bool(d, &["open_world_hint", "openWorldHint"])?,
-            deferred_hint: get_bool(d, &["deferred_hint", "deferredHint"])?,
-        };
-        Ok(())
-    }
-
-    /// Suggested follow-up tools for this declaration (issue #342).
-    ///
-    /// Returns ``None`` when neither ``on-success`` nor ``on-failure``
-    /// was declared. Otherwise returns a dict with string-list values
-    /// under the ``"on_success"`` / ``"on_failure"`` keys.
-    #[getter]
-    fn next_tools<'py>(
-        &self,
-        py: Python<'py>,
-    ) -> pyo3::PyResult<Option<pyo3::Bound<'py, pyo3::types::PyDict>>> {
-        if self.next_tools.on_success.is_empty() && self.next_tools.on_failure.is_empty() {
-            return Ok(None);
-        }
-        let d = pyo3::types::PyDict::new(py);
-        d.set_item("on_success", self.next_tools.on_success.clone())?;
-        d.set_item("on_failure", self.next_tools.on_failure.clone())?;
-        Ok(Some(d))
-    }
-
-    /// Set ``next_tools`` from a dict with optional ``on_success`` /
-    /// ``on_failure`` list-of-string values. Passing ``None`` clears
-    /// both lists.
-    #[setter]
-    fn set_next_tools(
-        &mut self,
-        value: Option<&pyo3::Bound<'_, pyo3::types::PyAny>>,
-    ) -> pyo3::PyResult<()> {
-        use pyo3::types::PyDict;
-        let Some(v) = value else {
-            self.next_tools = NextTools::default();
-            return Ok(());
-        };
-        if v.is_none() {
-            self.next_tools = NextTools::default();
-            return Ok(());
-        }
-        let dict = v.cast::<PyDict>().map_err(|_| {
-            pyo3::exceptions::PyTypeError::new_err(
-                "next_tools must be a dict with optional on_success/on_failure list keys, or None",
-            )
-        })?;
-        let on_success: Vec<String> = dict
-            .get_item("on_success")
-            .ok()
-            .flatten()
-            .map(|v| v.extract())
-            .transpose()?
-            .unwrap_or_default();
-        let on_failure: Vec<String> = dict
-            .get_item("on_failure")
-            .ok()
-            .flatten()
-            .map(|v| v.extract())
-            .transpose()?
-            .unwrap_or_default();
-        self.next_tools = NextTools {
-            on_success,
-            on_failure,
-        };
-        Ok(())
-    }
 }
 
 impl std::fmt::Display for ToolDeclaration {
